Bump gradle wrapper from 6.9 to 7.6.5, to work on latest XCode stack
This is needed because Bitrise CI's install-missing-android-tools
step uses Java 17 when run on the latest osx-xcode-16.4.x stack,
even when you ask it not to; Gradle 6.9 won't work with it. Gradle 7
supports both the Java 17 the NDK installer step requires, and the
Java 11 that the project itself requires.

The _right_ way to fix this is to migrate to RN 0.68.